requirements.txt
torch>=2.2.0
lightning>=2.2.0
pytorch-forecasting>=1.1.0

pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
requests>=2.31.0
tqdm>=4.66.0

folium>=0.16.0
streamlit-folium>=0.20.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
branca>=0.7.0

streamlit>=1.35.0
python-dotenv>=1.0.0
rich>=13.7.0

twilio>=8.12.0
earthengine-api>=0.1.390